+++ wgpu-core/src/device/trace/record.rs
@@ -866,7 +866,7 @@ fn action_to_owned(action: Action<'_, PointerReferences>) -> Action<'static, Poi
         A::DestroyExternalTexture(external_texture) => A::DestroyExternalTexture(external_texture),
         A::DestroySampler(sampler) => A::DestroySampler(sampler),
         A::GetSurfaceTexture { id, parent } => A::GetSurfaceTexture { id, parent },
-        A::Present(surface) => A::Present(surface),
+        A::Present(a, b) => A::Present(a, b),
         A::DiscardSurfaceTexture(surface) => A::DiscardSurfaceTexture(surface),
         A::DestroyBindGroupLayout(layout) => A::DestroyBindGroupLayout(layout),
         A::GetRenderPipelineBindGroupLayout {