+++ wgpu-hal/src/gles/queue.rs
@@ -1955,6 +1955,12 @@ impl crate::Queue for super::Queue {
     unsafe fn get_timestamp_period(&self) -> f32 {
         1.0
     }
+
+    unsafe fn wait_for_idle(&self) -> Result<(), crate::DeviceError> {
+        let gl = &self.shared.context.lock();
+        unsafe { gl.finish() };
+        Ok(())
+    }
 }
 
 #[cfg(send_sync)]